+++ src/app/sheeet/sheet.module.js
@@ -3,6 +3,6 @@
   // 工作表模块
 
   angular
-    .module('pf.sheet', ['pf.core', 'pf.table', 'pf.condition']);
+    .module('pf.sheet', ['pf.core', 'pf.user', 'pf.table', 'pf.condition']);
 
 })();